Fixed some usage of size_t.

+++ bm_gpu.hpp
@@ -8,7 +8,7 @@ namespace gpu {
 template <typename Tin, typename Tout>
 void LaplacianOfGaussian(Tin const *const src, Tout *const dst,
                          size_t src_pitch, size_t dst_pitch,
-                         size_t rows, size_t cols);
+                         int rows, int cols);
 
 }
